lab2/requirements.txt
langchain==0.3.7
langgraph==0.2.34
python-dotenv>=1.0
langchain-google-genai
